third_party/WebKit/LayoutTests/fast/dom/custom/lifecycle-created-innerHTML.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script>
description("This test ensures that the lifecycle callback of a parser-made element is visible in following script block.")

window.callbacksCalled = [];

function trackingReadyCallback()
{
    window.callbacksCalled.push(this.id);
}

document.registerElement("x-foo", { prototype: Object.create(HTMLElement.prototype, { createdCallback: { value: trackingReadyCallback } }) });
document.registerElement("x-bar", { extends: "div", prototype: Object.create(HTMLDivElement.prototype, { createdCallback: { value: trackingReadyCallback } }) });
document.registerElement("x-baz", { prototype: Object.create(HTMLElement.prototype, { createdCallback: { value: trackingReadyCallback } }) });
</script>
</head>
<body>
<div id="container">
<script>
// Testing element siblings
window.callbacksCalled = [];
container.innerHTML = "<x-foo id='a'></x-foo><div id='b' is='x-bar'></div><x-baz id='c'></x-baz>";
shouldBe("window.callbacksCalled", "['a', 'b', 'c']");

// Testing nested elements
window.callbacksCalled = [];
container.innerHTML = "<x-foo id='a'><div id='b' is='x-bar'><x-baz id='c'></x-baz></div></x-foo>";
shouldBe("window.callbacksCalled", "['a', 'b', 'c']");
</script>
</body>
</html>
