src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/custom/lifecycle-created-innerHTML-expected.txt

CONSOLE ERROR: line 15: Uncaught TypeError: document.registerElement is not a function
This test ensures that the lifecycle callback of a parser-made element is visible in following script block.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL window.callbacksCalled should be a,b,c. Was .
FAIL window.callbacksCalled should be a,b,c. Was .
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
